+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.0] - 2024-09-08
+
 ### Added
 
 - Icon for Database tool
@@ -154,7 +156,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
-[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.4.3...HEAD
+[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.4.3...v1.5.0
 [1.4.3]: https://github.com/pandier/intellij-discord-rp/compare/v1.4.2...v1.4.3
 [1.4.2]: https://github.com/pandier/intellij-discord-rp/compare/v1.4.1...v1.4.2
 [1.4.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.4.0...v1.4.1